+++ src/lib/src/client/index.ts
@@ -55,6 +55,8 @@ export type TableParams<Row> = {
     selectBy?: Field<Row>,
     highlight?: boolean,
     i18n?: Internationalization,
+    recursiveFilter?: boolean,
+    recursiveSearch?: boolean,
 }
 
 export type Criterion = {